virtual_modems/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pycryptodome>=3.20.0
httpx>=0.27.0
